moves deepGet and deepClone to src/app/helpers/json
