+++ data/preprocessor.js
@@ -18,9 +18,3 @@ const Preprocessor = (jsonRaw = {}, rules = defaultRules) => {
 };
 
 export default Preprocessor;
-
-// taken from https://medium.com/javascript-inside/safely-accessing-deeply-nested-values-in-javascript-99bf72a0855a
-export const deepGet = (path, object) =>
-  path.reduce((xs, x) => (xs && xs[x] ? xs[x] : null), object);
-
-export const deepClone = originalObj => JSON.parse(JSON.stringify(originalObj));
